python-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
opencv-python-headless==4.9.0.80
mediapipe==0.10.33
pika==1.3.2
python-dotenv==1.0.1
numpy>=2.3.2